refactor: collapse duplicate arb-planner component trees into shared core
The authed `/arbitration-planner` and public `/arb-planner-public` routes
maintained parallel, drifting copies of their client, roster section, and
plan-comparison components. Roster-rendering fixes had to be made twice.

Extract a single generic component tree under `web/components/arb-planner/`:

- `ArbPlannerCore<T>` — shared client holding all tab/plan-CRUD/validation/
  budget-tracking state and rendering. Generic over `T extends ArbPlannerPlayer`.
- `TeamRosterSection<T>` — collapsible per-team roster table. Extra columns are
  prop-driven (`showSurplus`, `adjustedSurplus`, `hoverDataMap`, `nameMode`),
  defaulting to the read-only/public behavior.
- `PlanComparison<T>` — comparison table whose middle metric column is supplied
  via a `metricColumn` config (authed surplus vs public PPG).
- `types.ts` — `ArbPlannerPlayer` base shape satisfied by both
  `ArbitrationTarget` and `PublicArbPlayer`, with compile-time assertions.

`ArbPlannerClient` and `PublicArbPlannerClient` become thin wrappers that pass
mode-specific props. The old per-route component files become deprecated
re-export shims (kept to avoid breaking stray imports). No visible UI or
behavior changes in either route: public stays read-only without surplus/hover,
authed keeps editing, hover cards, adjusted surplus, and "Create from Suggested".

Adds a test suite for the shared `TeamRosterSection` covering the prop-driven
column visibility and allocation clamping. Full web suite: 229 tests pass;
`just typecheck` and `just lint` clean.
